+++ src/functions/online_ops.py
@@ -5,19 +5,19 @@ from email.message import EmailMessage
 import smtplib
 from decouple import config
 
-NEWS_API_KEY = config("NEWS_API_KEY")
-OPENWEATHER_APP_ID = config("OPENWEATHER_APP_ID")
-TMDB_API_KEY = config("TMDB_API_KEY")
-EMAIL_ACCOUNT = config("EMAIL_ACCOUNT")
-EMAIL_PASSWORD = config("EMAIL_PASSWORD")
+# NEWS_API_KEY = config("NEWS_API_KEY")
+# OPENWEATHER_APP_ID = config("OPENWEATHER_APP_ID")
+# TMDB_API_KEY = config("TMDB_API_KEY")
+# EMAIL_ACCOUNT = config("EMAIL_ACCOUNT")
+# EMAIL_PASSWORD = config("EMAIL_PASSWORD")
 
 
 
-def get_weather_report(city):
-    url = f"http://api.openweathermap.org/data/2.5/weather?q={city}&appid={OPENWEATHER_APP_ID}&units=imperial"
-    res = requests.get(url).json()
-    weather = res["weather"][0]["main"]
-    temperature = res["main"]["temp"]
-    feels_like = res["main"]["feels_like"]
-    return weather, f"{temperature}°F", f"{feels_like}°F" # alt and 0176 on right keypad = degree symbol
+# def get_weather_report(city):
+#     url = f"http://api.openweathermap.org/data/2.5/weather?q={city}&appid={OPENWEATHER_APP_ID}&units=imperial"
+#     res = requests.get(url).json()
+#     weather = res["weather"][0]["main"]
+#     temperature = res["main"]["temp"]
+#     feels_like = res["main"]["feels_like"]
+#     return weather, f"{temperature}°F", f"{feels_like}°F" # alt and 0176 on right keypad = degree symbol
